move reroute and select route segments to the general view
