+++ src/groups/RouteNetworkDiagram/PassageView/PassageView.scss
@@ -16,14 +16,14 @@
 .passage-view_title {
   font-weight: bold;
   padding-right: 50px;
-  font-size: 14px;
+  font-size: 16px;
 }
 
 .passage-view-row-title-header {
   display: flex;
   width: 100%;
   justify-content: space-between;
-  padding: 5px 10px;
+  padding: 10px;
   align-items: center;
 }
